Use the nodejs buildpack to "compile" the app
The compile step installs node and npm and dependencies from the package.json.

Because node (etc) is placed in a non-standard location, we adjust the path to allow the call to npm to work. And it does.

I haven't got it working with redis yet. In theory, when it comes up with docker-compose, the redis container should be reachable at the redis hostname.